==> 2_Data Prep_death_records_public.ipynb <==
### Capstone 1 - Washington state linkage of infant death, birth, and mother's hospitalization discharge data

##### Maya Bhat-Gregerson

January 7, 2020

### B. PREPARATION OF DEATH DATA 2016-2018

### I. Data acquisition

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import pyodbc

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

I use SQL queries to get the death and birth variables I am likely to need for linking the records. 

NOTE: I renamed all variables from death records so that they begin with 'd' to distinguish the fields from those in birth records with the same names.

**DEATH DATA 2016, 2017, 2018**

## CONNECT TO WHALES & USE SQL QUERY FOR DEATH DATA SET

driver = '{SQL Server Native Client 11.0}'

conn = pyodbc.connect(
        Trusted_Connection='Yes',
        Driver='{ODBC Driver 13 for SQL Server}',
        Server='###',
        Database='###'
        )

querystring = ("SELECT SFN_NUM as 'dsfn'," + 
        "BIRTH_SFN_NUM as 'dbirsfn', " + 
        "SSN as 'dssn', " +
        "ISNULL(GNAME, 'NaN') as 'dfname', " +
        "ISNULL(MNAME, 'NaN') as 'dmname', " +
        "ISNULL(LNAME, 'NaN') as 'dlname', " +
        "ISNULL(MOTHER_GNAME, 'NaN') as 'dmom_fname', " +
        "ISNULL(MOTHER_MNAME, 'NaN') as 'dmom_mname', " + 
        "ISNULL(MOTHER_LNAME, 'NaN') as 'dmom_maiden', " +
        "ISNULL(FATHER_LNAME, 'NaN') as 'ddad_lname', " +
        "ISNULL(FATHER_MNAME, 'NaN') as 'ddad_mname', " +    
        "ISNULL(FATHER_GNAME, 'NaN') as 'ddad_fname', " +           
        "ISNULL(SEX, 'NaN') as 'dsex', " + 
		"DOB as 'ddob', " + 
		"ISNULL(SUBSTRING(DOB, 1,2), '99') as 'ddobm', " + 
		"ISNULL(SUBSTRING(DOB, 4,2), '99') as 'ddobd', " +
		"ISNULL(SUBSTRING(DOB, 7,4), '9999') as 'ddoby', " + 
		"ISNULL(DOD, '  ') as 'ddod'," + 
		"ISNULL(SUBSTRING(DOD, 1,2), '99') as 'ddodm', " + 
		"ISNULL(SUBSTRING(DOD, 4,2), '99') as 'ddodd', " + 
		"ISNULL(SUBSTRING(DOD, 7,4), '9999') as 'ddody', " +
		"ISNULL(BPLACE_ST_FIPS_CD, '  ') as 'dbirplstatefips', " + 
		"ISNULL(BPLACE_CNT, '  ') as 'dbircountryl'," + 
		"ISNULL(DNAME_CITY, '  ') as 'ddthcityl', " + 
		"ISNULL(DCOUNTY, '  ') as 'ddthcountyl'," + 
		"ISNULL(DSTATEL, '  ') as 'ddthstatel', " + 
		"ISNULL(SUBSTRING(DZIP9, 1,5), '99999') as 'ddthzip', " + 
		"ISNULL(RES_CITY, '  ') as 'drescity', " + 
		"RIGHT('00000' + ISNULL(RES_CITY_FIPS_CD, '99999'), 5) as 'drescityfips', " + 
		"ISNULL(RES_COUNTY, '  ') as 'drescountyl', " + 
		"RIGHT('000' + ISNULL(RES_COUNTY_FIPS_CD, '999'), 3) as 'drescntyfips', " + 
		"ISNULL(RES_STATE_FIPS_CD, '  ') as 'dresstatefips', " + 
		"ISNULL(SUBSTRING(RES_ZIP, 1,5), '99999') as 'dreszip'" + 
"FROM [wa_vrvweb_events].[VRV_DEATH_TBL]" +
"WHERE ((DATE_DEATH_YEAR = 2016) OR (DATE_DEATH_YEAR = 2017) OR (DATE_DEATH_YEAR = 2018))" +
	"AND VRV_REGISTERED_FLAG = '1'" +
	"AND FL_CURRENT = '1'" +
	"AND FL_VOIDED = '0'")

dth1618= pd.read_sql_query(querystring, conn)

## SAVE DATA AS CSV FILE

dth1618.to_csv(r'###\Data\dth1618_raw.csv', index=None, header=True)

d1618 = pd.read_csv(r'###\Data\dth1618_raw.csv', low_memory=False)

d1618.shape

d1618.info()

### II. Data cleaning and standardization

# look at data
#d1618.head()

  - Check death record number 'dsfn' and convert to integer if necessary

d1618.dsfn.describe()

dth1618['dsfn'] = dth1618['dsfn'].astype(int)

dth1618.dsfn.dtype

**RESTRICT TO WA RESIDENTS**

d1618['dresstatefips'].value_counts(dropna=False)

# keep only WA residents and check if successful in filtering out other state residents.
d1618 = d1618[(d1618['dresstatefips']=="WA")]
d1618['dresstatefips'].value_counts(dropna=False)

#### CHECK FOR NULL VALUES

d1618.isna().sum()

Probably will not use mothers' and infants' middle name for linking as there are too many missing values.

#### STANDARDIZE ALL STRING VARIABLES

First, middle, and last names of infants and mothers as well as city names will be standardized by converting these columns to upper case text, removing white spaces, removing hyphens and other punctuation marks.

# convert to upper case
d1618 = d1618.apply(lambda x: x.str.upper() if type(x) == str else x)

#remove white spaces, punctuation
d1618 = d1618.apply(lambda x: x.str.strip() if type(x) == str else x)
d1618 = d1618.applymap(lambda x: x.replace(" ", "") if type(x) == str else x)
d1618 = d1618.applymap(lambda x: x.replace("-", "") if type(x) == str else x)
d1618 = d1618.applymap(lambda x: x.replace(".", "") if type(x) == str else x)
d1618 = d1618.applymap(lambda x: x.replace("'", "") if type(x) == str else x)

#d1618.tail(30)

- Verified that all string transformations were successful.

#### CHECK FOR OUT OF RANGE VALUES

#create dictionary of valid values so that each variable can be checked to make sure there is no
# out of range value.

valids = {'sex': ['M', 'F', 'U'],
          'dobm': [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 99],
          'dodm': [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 99],
          'dobd': np.r_[1:32 ,99],
          'dodd': np.r_[1:32 ,99],
          'doby': list(range(1900,2019)),
          'dody': [2016,2017,2018],
         'rcntyfips': np.r_[list(range(1, 78, 2)), 99],
         'certtype': ['R'],
         'birthstatefips': ['WA'], 
         'rstatefips': ['WA']}

# check for out of range values for 'dsex'

chksex = d1618['dsex'].isin(valids['sex'])
len(d1618[~chksex])

# check for out of range values for 'ddobm'

chkdobm = d1618['ddobm'].isin(valids['dobm'])
len(d1618[~chkdobm])

# check for out of range values for 'ddody'

chkdody = d1618['ddody'].isin(valids['dody'])
len(d1618[~chkdody])

d1618['ddody'].value_counts(dropna=False)

# check for out of range values for 'ddoby'

chkdoby = d1618['ddoby'].isin(valids['doby'])
len(d1618[~chkdoby])

d1618.loc[~chkdoby].ddoby.value_counts(dropna=False)


dobyerrors = d1618[~chkdoby][['ddob']]

dobyerrors

- unable to correct dob year without additional information on date of birth.

# check for out of range values for 'dobd'

chkdobd = d1618['ddobd'].isin(valids['dobd'])
len(d1618[~chkdobd])


# check for out of range values for 'drescntyfips'

chkrcounty = d1618['drescntyfips'].isin(valids['rcntyfips'])
len(d1618['drescntyfips'][~chkrcounty])

rcntyerrors = d1618[~chkrcounty][['dsfn','drescntyfips', 'drescountyl','drescity']]

rcntyerrors

 - cannot replace with correct values without any information on residence city or county.

###### Format date variables as dates

d1618.ddod = pd.to_datetime(d1618.ddod)


d1618.to_csv(r'###\Data\d1618_clean.csv', index=None, header=True)

Next step - clean/prepare the manually linked labeled data for infant birth and death linked file.
