==> 4_Data_Prep_for_classification_public.ipynb <==
### Record Linkage: preparing data sets to train classifiers

After acquiring birth and death data for 2016 through 2018 for Washington state from the Microsoft SQL Server database, the next steps are to: <br>
 <br> 
(1) pare down the data sets to only the variables needed for classification, <br> 
(2) preprocess data sets in preparation for classification, <br> 
(3) create candidate pairs of birth and death records, <br>
(4) compare select variables for each candidate pair of records to compute similarity scores on a 0 to 1 scale, <br>
(5) use the linked infant birth-death records for 2016-17 as the "golden records" data set to label candidate pairs as matches
    and non-matches, <br>
(6) Split the data into training and testing data sets. <br>
 <br>
At this point the data are ready to use in building classifiers.

import pandas as pd
import numpy as np
import recordlinkage as rl

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

#### (1) READ IN AND PARE DOWN DATA SETS TO LINKAGE VARIABLES

Data sets used: birth 2016-18, death 2016-18, linked infant death 2016-17.  

Birth and death data are restricted to events that occurred in 2016 and 2017; I will use 2018 to validate the model after training and testing on 2016 and 2017 combined.

All data are limited to births or deaths that occurred in Washington state to in state residents.

Variables to be used for classifying record pairs as matches and non-matches include infants' (decedents'):
- first and last names, 
- mother's first and last names, 
- father's first and last names, 
- sex, 
- date of birth month, day, and year (as separate variables), 
- residence city and county at the time of death.

These variables are preserved for both the birth and death data sets. 

Birth and death certificate numbers were also included in order to create unique id's for each record pair.

##### Death data

d1618 = pd.read_csv(r'###\Data\d1618_clean.csv', 
                    low_memory = False)

d1617 = d1618[(d1618['ddody']!=2018)]
d1617 = d1617[(d1617.ddthstatel=='WASHINGTON')]

d1617 = d1617.loc[:, ['dsfn', 'dfname', 'dlname', 'dmom_fname','dmom_maiden','ddad_fname', 
                      'ddad_lname','dsex','ddobm', 'ddobd', 'ddoby','drescity','drescountyl']]

d1617.info()

##### Birth data

b1618 = pd.read_csv(r'###\Data\b1618_clean.csv', 
                    low_memory = False)

b1617 = b1618[(b1618.bdoby !=2018)]

b1617.bdoby.value_counts(dropna=False)

# check to make sure only Washington state births are included

b1617.bbirplstatefips.value_counts(dropna=False)

# check to make sure only Washington state residents are included

b1617.b_momresstatefips.value_counts(dropna=False)

b1617 = b1617.loc[:, ['bsfn','bfname','blname', 'bmom_fname', 'bmom_lname', 'bdad_lname', 
                      'bdad_fname','bsex', 'bdobm', 'bdobd', 'bdoby', 'b_momrescity', 
                      'b_momrescountyl']]

b1617.info()

#####  "Golden pairs" - linked infant birth-death file 2016-2017

The records in the 2016-17 infant birth-death file were manually linked birth and death records that represent the true matches for infant deaths occurring in those years.

linked1617 = pd.read_csv(r'###\WA1617infantDeath_dthdata.csv')

linked1617.info()

linked1617.ddthstatel.value_counts(dropna=False)

linked1617.dbirplstatefips.value_counts(dropna=False)

# keep only birth and death certificate numbers of true matches to label the training data set

linked1617 = linked1617.loc[:,['lbsfn','ldsfn']]

linked1617.info()

linked1617.to_csv(r'###\data\clean\linked1617_labels.csv')

#### (2) Preprocessing: data cleaning and standardizing

###### DEATH

from recordlinkage.preprocessing import clean

d1617['dfname_clean'] = clean(d1617.dfname,
                              lowercase = True,
                              replace_by_none = '[\s-]+',
                              strip_accents = 'unicode'
                             )

d1617['dlname_clean'] = clean(d1617.dlname,
                              lowercase = True,
                              replace_by_none = '[\s-]+',
                              strip_accents = 'unicode'
                             )

d1617['dmom_fname_clean'] = clean(d1617.dmom_fname,
                                 lowercase = True,
                                  replace_by_none = '[\s-]+',
                                  strip_accents = 'unicode'
                                 )

d1617['dmom_lname_clean'] = clean(d1617.dmom_lname,
                                 lowercase = True,
                                  replace_by_none = '[\s-]+',
                                  strip_accents = 'unicode'
                                 )

d1617['ddad_fname_clean'] = clean(d1617.ddad_fname,
                                lowercase = True,
                                  replace_by_none = '[\s-]+',
                                  strip_accents = 'unicode'
                                 )

d1617['ddad_lname_clean'] = clean(d1617.ddad_lname,
                                lowercase = True,
                                  replace_by_none = '[\s-]+',
                                  strip_accents = 'unicode'
                                 )

###### Create phonetic encoding of infant, mother, and father names

d1617['dmom_fullname'] =  d1617.dmom_fname_clean + " " +  d1617.dmom_lname_clean

d1617['ddad_fullname'] =  d1617.ddad_fname_clean + " " +  d1617.ddad_lname_clean

d1617['dinf_fullname'] =  d1617.dfname_clean + " " +  d1617.dlname_clean

d1617['dmom_phon'] = rl.preprocessing.phonetic(d1617['dmom_fullname'], method = 'metaphone')
d1617['ddad_phon'] = rl.preprocessing.phonetic(d1617['ddad_fullname'], method = 'metaphone')
d1617['dinf_phon'] = rl.preprocessing.phonetic(d1617['dinf_fullname'], method = 'metaphone')

# View sample of death records to make sure the cleaning/standardization steps did what they were supposed to do.

d1617.head()

##### BIRTH

b1617['bfname_clean'] = clean(b1617.bfname,
                              lowercase = True,
                              replace_by_none = '[\s-]+',
                              strip_accents = 'unicode'
                             )

b1617['blname_clean'] = clean(b1617.blname,
                              lowercase = True,
                              replace_by_none = '[\s-]+',
                              strip_accents = 'unicode'
                             )

b1617['bmom_fname_clean'] = clean(b1617.bmom_fname,
                              lowercase = True,
                              replace_by_none = '[\s-]+',
                              strip_accents = 'unicode'
                             )

b1617['bmom_lname_clean'] = clean(b1617.bmom_lname,
                                 lowercase = True,
                                  replace_by_none = '[\s-]+',
                                  strip_accents = 'unicode'
                                 )

b1617['bdad_fname_clean'] = clean(b1617.bdad_fname,
                                lowercase = True,
                                  replace_by_none = '[\s-]+',
                                  strip_accents = 'unicode'
                                 )

b1617['bdad_lname_clean'] = clean(b1617.bdad_lname,
                                lowercase = True,
                                  replace_by_none = '[\s-]+',
                                  strip_accents = 'unicode'
                                 )

b1617['bmom_fullname'] =  b1617.bmom_fname_clean + " " +  b1617.bmom_lname_clean
b1617['bdad_fullname'] =  b1617.bdad_fname_clean + " " +  b1617.bdad_lname_clean
b1617['binf_fullname'] =  b1617.bfname_clean + " " +  b1617.blname_clean

b1617['bmom_phon'] = rl.preprocessing.phonetic(b1617['bmom_fullname'], method = 'metaphone')
b1617['bdad_phon'] = rl.preprocessing.phonetic(b1617['bdad_fullname'], method = 'metaphone')
b1617['binf_phon'] = rl.preprocessing.phonetic(b1617['binf_fullname'], method = 'metaphone')

## Check sample of birth data set to make sure cleaning/standardization worked

#b1617.head()

b1617.to_csv(r'###\data\clean\b1617_clean.csv', index=None, header=True)
d1617.to_csv(r'###\data\clean\d1617_clean.csv', index=None, header=True)

##### Reduction of birth and death data size

The two primary reasons for reducing the numbers of birth and death records are (1) to address class imbalance between matched and unmatched records, and (2) to reduce the computational load associated with creating candidate pairs of all birth records (n = 259,908) and all death records (n= 168,701)  for this two year period.  

As the purpose of this project is to create a linked birth and death record file for infants (decedents who died before age 1 year) the death data set can be reduced by excluding records for decedents who were older than 364 days at the time of death. This reduces the number of deaths 


For the birth data set, I undersampled the birth records that will not have matches in the death data (by comparing with the exising "golden pairs" i.e. true matches) so that the final birth data set contains roughly equal number of records that have matches in the reduced death data set and those that don't have matches.

To implement this reduction, the first step is to use the linked1617 golden pairs and use their ID numbers to identify records that will match or not match in the birth and death data sets.

#b1617 = pd.read_csv(r'###\data\clean\b1617_clean.csv')
#d1617 = pd.read_csv(r'###\data\clean\d1617_clean.csv')
#linked1617 = pd.read_csv(r'###\data\clean\linked1617_labels.csv')

matched_dsfn = linked1617['ldsfn'].tolist()
len(matched_dsfn)

matched_bsfn = linked1617['lbsfn'].tolist()
len(matched_bsfn)

# separate birth records without matches in death data

b1617_not_linked = b1617[(~b1617['bsfn'].isin(matched_bsfn))]

# Select a random sample of 650 unmatched birth records

b1617_not_linked_undersample = b1617_not_linked.sample(n=650, random_state=42)

# separate birth records with matches in death data set

b1617_linked =  b1617[b1617['bsfn'].isin(matched_bsfn)]

# create a new birth data set consisting of the birth records that have matches and the sample 
#of 650 records that do not have matches in the death data.

b1617_fin = pd.concat([b1617_not_linked_undersample, b1617_linked], axis=0)

# as both birth and death record IDs in WA are 10 digits long and begin with the 4 digit year of the
# event, I make sure that there are no common (non-unique) numbers between the birth and death data
# by adding numbers to the ID integers so that birth IDs start with 1 and death IDs start with 5.

# set index for birth records to the birth record number ('bsfn')

b1617_fin.bsfn = b1617_fin.bsfn + 10000000000
b1617_fin = b1617_fin.set_index('bsfn')

# repeating the process above to keep only death records that are also present in the true match
#linked data set.

d1617_linked =  d1617[d1617['dsfn'].isin(matched_dsfn)]

d1617_fin = d1617_linked.copy()

d1617_fin.dsfn = d1617_fin.dsfn + 50000000000

d1617_fin = d1617_fin.set_index('dsfn')

b1617_fin.info()

d1617_fin.info()

# change record IDs so that they are unique.

linked1617.lbsfn = linked1617.lbsfn + 10000000000
linked1617.ldsfn = linked1617.ldsfn + 50000000000

linked1617.to_csv(r'###data\clean\linked1617_labels.csv')

#### BLOCKING AND INDEXING CANDIDATE RECORD PAIRS

To reduce the number of candidate pairs that will be compared on linking variables, the birth and death records in the final data sets are grouped into 'blocks' based on year of birth, month of birth, day of birth, and mother's name phonetically encoded. Each block will contain only those records that have the same values for these fields.

#initalize indexer and specify blocking variables

indexer = rl.Index()
indexer.block(left_on = 'bdoby', right_on = 'ddoby')
indexer.block(left_on = 'bdobm', right_on = 'ddobm' )
indexer.block(left_on = 'bdobd', right_on = 'ddobd')
indexer.block(left_on = 'bmom_phon', right_on = 'dmom_phon')

# create candidate pairs between reduced birth and death data sets

candidate_pairs = indexer.index(b1617_fin, d1617_fin)

'''The following is the number of candidate_pairs created by pairing all birth and death records in 
the reduced data sets i.e. 631 death X 1,281 birth records.
'''
len(candidate_pairs)

##### CANDIDATE PAIR COMPARISONS

The candidate pairs created above are compared on similarities of the infants' first and last names, mothers' last name, the phonetic encoding of mothers' full name, fathers' last name, the infants' sex, and the mothers' county of residence at the time of birth and death.

# initiate comparison function

c=rl.Compare()

c.string('blname_clean', 'dlname_clean', method = 'jarowinkler', label = 'cmp_inf_lname')
c.string('binf_phon', 'dinf_phon', method = 'jarowinkler', label = 'cmp_inf_phonetic')
c.string('bmom_lname_clean', 'dmom_lname_clean', method = 'jarowinkler', label = 'cmp_mom_lname')
c.string('bmom_phon', 'dmom_phon', method = 'jarowinkler', label = 'cmp_mom_phonetic')
c.string('bdad_lname_clean', 'ddad_lname_clean', method = 'jarowinkler', label = 'cmp_dad_lname')
c.string('bdad_phon', 'ddad_phon', method = 'jarowinkler', label = 'cmp_dad_phonetic')
c.numeric('bdobm', 'ddobm', method = 'linear', scale = 1, label = 'cmp_dobm')
c.numeric('bdoby', 'ddoby', method = 'linear', scale = 1, label = 'cmp_doby')
c.numeric('bdobd', 'ddobd', method = 'linear', scale = 1, label = 'cmp_dobd')
c.exact('bsex', 'dsex', label = 'cmp_sex')
c.exact('b_momrescountyl', 'drescountyl', label = 'cmp_rescounty')

# Apply comparison to the candidate pairs created above.
# The resulting data frame contains the comparision scores (from 0 to 1) of the variables above.
# True matches will be closer to 1 on these variables.

comp_scores = c.compute(candidate_pairs, b1617_fin, d1617_fin)

len(comp_scores), type(comp_scores)

comp_scores.info()

comp_scores.tail()

**Add variable to indicate whether record pairs in comp_scores are matches or non-matches**

The golden pairs (true match) linked data set allows labeling of the entire comp_scores data set.

# set a multiIndex for the true match record ID pairs

labels = linked1617.set_index(['lbsfn', 'ldsfn'])

# add a variable match (set to 1) indicating that the record pairs in dataframe 'labels' are all matches
labels['match'] = 1

# concatenate the comparison scores data with the labels and fill all blank values for 'match' variables with 0 indicating
# no match.

comp_scores_m = (pd.concat([comp_scores, labels], axis=1)
       .fillna(0)
       .astype(int))

comp_scores_m.info()

comp_scores_m.to_csv(r'###\data\clean\infdth_comparison_1617_labelled.csv', header=True)

The comparison score data frame can now be used in the next step when building classifiers.
